restructure loading/initializatino and manager
