+++ corebehrt/modules/setup/directory.py
@@ -6,7 +6,7 @@ from os.path import basename, join, splitext
 from shutil import copyfile, rmtree
 
 from corebehrt.common.checks import check_categories
-from corebehrt.common.config import Config, load_config
+from corebehrt.modules.setup.config import Config, load_config
 
 logger = logging.getLogger(__name__)  # Get the logger for this module
